Refactor formatters
