+++ lib/dry/logger.rb
@@ -4,12 +4,29 @@ require "logger"
 require "pathname"
 
 require "dry/logger/level"
+
 require "dry/logger/backends/io"
 require "dry/logger/backends/file"
+
+require "dry/logger/formatters/string"
+require "dry/logger/formatters/json"
+require "dry/logger/formatters/application"
 
 module Dry
   module Logger
     include Level
+
+    def self.formatters
+      @formatters ||= {}
+    end
+
+    def self.register_formatter(name, formatter)
+      formatters[name] = formatter
+    end
+
+    register_formatter(:string, Formatters::String)
+    register_formatter(:json, Formatters::JSON)
+    register_formatter(:application, Formatters::Application)
 
     def self.new(stream: $stdout, **opts)
       backend =
@@ -20,7 +37,19 @@ module Dry
           raise ArgumentError, "unsupported stream type #{stream.class}"
         end
 
-      backend.new(stream: stream, **opts)
+      formatter_opt = opts[:formatter]
+
+      formatter =
+        case formatter_opt
+        when Symbol then formatters.fetch(formatter_opt).new(**opts)
+        when Class then formatter_opt.new(**opts)
+        when NilClass then formatters[:string].new(**opts)
+        when ::Logger::Formatter then formatter_opt
+        else
+          raise ArgumentError, "unsupported formatter option #{formatter_opt.inspect}"
+        end
+
+      backend.new(stream: stream, **opts, formatter: formatter)
     end
   end
 end
